+++ packages/webapp/src/_app/eos/ual/softkey/hooks.ts
@@ -1,6 +1,7 @@
 import { putEncryptionKey } from "encryption";
 import { actionShowUALSoftkeyModal, useGlobalStore } from "_app";
 import { actionSetEncryptionPassword } from "_app/actions";
+import { MODAL_INDEX } from "_app/ui";
 
 export interface UALSoftKeyLoginHook {
     isOpen: boolean;
@@ -9,17 +10,20 @@ export interface UALSoftKeyLoginHook {
     updateEncryptionPassword: (publicKey: string, privateKey: string) => void;
 }
 
+const UAL_MODAL_BACKGROUND_INDEX = `${MODAL_INDEX - 1}`;
+const UAL_MODAL_ORIGINAL = "2147483647";
+
 export const useUALSoftkeyLogin = (): UALSoftKeyLoginHook => {
     const { state, dispatch } = useGlobalStore();
     const { ualSoftkeyModal } = state;
     const { isOpen, resolver } = ualSoftkeyModal;
 
     const show = () => {
-        // hack to send ual box behind
+        // hack to send ual box behind eden modals
         const ualBox = document.getElementById("ual-box");
         if (ualBox && ualBox.parentElement) {
             const ualBoxModalOverlay = ualBox.parentElement;
-            ualBoxModalOverlay.style.zIndex = "49"; // send behind our modal
+            ualBoxModalOverlay.style.zIndex = UAL_MODAL_BACKGROUND_INDEX; // send behind our modal
         }
 
         return new Promise<string>((resolve) => {
@@ -36,7 +40,7 @@ export const useUALSoftkeyLogin = (): UALSoftKeyLoginHook => {
         const ualBox = document.getElementById("ual-box");
         if (ualBox && ualBox.parentElement) {
             const ualBoxModalOverlay = ualBox.parentElement;
-            ualBoxModalOverlay.style.zIndex = "2147483647"; // original ual z-index
+            ualBoxModalOverlay.style.zIndex = UAL_MODAL_ORIGINAL; // original ual z-index
         }
     };
