+++ packages/webapp/src/pages/_app.tsx
@@ -19,7 +19,7 @@ import {
 import { EdenUALProvider, Store, Toaster } from "_app";
 
 import EncryptionPasswordModals from "encryption/components/encryption-password-modals";
-import { UalSoftKeyModals } from "_app/utils/softkey-ual-authenticator";
+import { UalSoftkeyLoginModal } from "_app/eos/ual/softkey";
 
 import "tailwindcss/tailwind.css";
 import "_app/styles/nprogress.tailwind.css";
@@ -64,7 +64,7 @@ const WebApp = ({ Component, pageProps }: AppProps) => {
                     <Hydrate state={pageProps.dehydratedState}>
                         <EdenUALProvider>
                             <EncryptionPasswordModals />
-                            <UalSoftKeyModals />
+                            <UalSoftkeyLoginModal />
                             <Component {...pageProps} />
                         </EdenUALProvider>
                     </Hydrate>
